Amazing refactoring stage 1. Add security for moderators.

+++ hr-api/src/main/java/com/digdes/controllers/DepartmentApiController.java
@@ -9,6 +9,7 @@ import com.digdes.exceptions.EntityNotFoundException;
 import com.digdes.exceptions.EntityUpdateException;
 import com.digdes.services.DepartmentDataService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
@@ -19,12 +20,8 @@ import java.util.List;
 @RestController
 @RequestMapping(path = "/departments")
 public class DepartmentApiController {
-//todo: constraints for bad request
-//todo: exceptions
     @Autowired
     private DepartmentDataService dataService;
-
-
 
     @PostMapping(path = "/create")
     public ResponseEntity<?> create(@RequestBody DepartmentDto info, Authentication authentication) {
@@ -32,9 +29,12 @@ public class DepartmentApiController {
             return ResponseEntity.badRequest().body("Department id must be null for create. Use update method");
         if (info.getName() == null || info.getTypeId() == null)
             return ResponseEntity.badRequest().body("Name and type id fields cannot be null");
-        try{
-            DepartmentResponseDto result = dataService.create(info);
-            return ResponseEntity.ok(result);
+        try {
+            if (dataService.checkUser(authentication.getName(), info.getParentId())) {
+                DepartmentResponseDto result = dataService.create(info);
+                return ResponseEntity.ok(result);
+            }
+           return ResponseEntity.status(HttpStatus.FORBIDDEN).body(String.format("User %s cannot moderate this department", authentication.getName()));
         }
         catch (EntityNotFoundException | EntityCreateException e){
             return ResponseEntity.internalServerError().body(e.getMessage());
@@ -42,31 +42,34 @@ public class DepartmentApiController {
     }
 
     @PutMapping(path = "/update")
-    public ResponseEntity<?> update(@RequestBody DepartmentDto info) {
+    public ResponseEntity<?> update(@RequestBody DepartmentDto info, Authentication authentication) {
         if (info.getId() == null)
             return ResponseEntity.badRequest().body("Department id must not be null for update. Use create method");
         if (info.getName() == null && info.getParentId() == null
-                && info.getHeadId() == null && info.getTypeId() == null
-                && info.getModeratorId() == null)
+                && info.getTypeId() == null && info.getModeratorId() == null)
             return ResponseEntity.badRequest().body("One of updating fields must not be null");
         try{
-            DepartmentResponseDto result = dataService.update(info);
-            return ResponseEntity.ok(result);
+            if (dataService.checkUser(authentication.getName(), info.getParentId())) {
+                DepartmentResponseDto result = dataService.update(info);
+                return ResponseEntity.ok(result);
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(String.format("User %s cannot moderate this department", authentication.getName()));
         }
         catch (EntityNotFoundException | EntityUpdateException e){
             return ResponseEntity.internalServerError().body(e.getMessage());
         }
     }
 
-
-    //todo: test delete trigger from path
     @DeleteMapping(path = "/delete")
-    public ResponseEntity<?> delete(@RequestBody DepartmentDto info) {
+    public ResponseEntity<?> delete(@RequestBody DepartmentDto info, Authentication authentication) {
         if (info.getId() == null)
             return ResponseEntity.badRequest().body("Department id must not be null for delete");
         try {
-            Boolean result = dataService.delete(info);
-            return ResponseEntity.ok(result);
+            if (dataService.checkUser(authentication.getName(), info.getParentId())) {
+                Boolean result = dataService.delete(info);
+                return ResponseEntity.ok(result);
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(String.format("User %s cannot moderate this department", authentication.getName()));
         }
         catch (EntityNotFoundException | EntityDeleteException e){
             return ResponseEntity.internalServerError().body(e.getMessage());
@@ -75,8 +78,7 @@ public class DepartmentApiController {
 
     @PostMapping(path = "/find")
     public ResponseEntity<?> find(@RequestBody DepartmentDto searchRequest) {
-        if (searchRequest.getName() == null && searchRequest.getParentId() == null
-                && searchRequest.getHeadId() == null && searchRequest.getTypeId() == null
+        if (searchRequest.getName() == null && searchRequest.getParentId() == null && searchRequest.getTypeId() == null
                 && searchRequest.getId() == null && searchRequest.getModeratorId() == null)
             return ResponseEntity.badRequest().body("One of fields must not be null");
         try {
@@ -92,11 +94,6 @@ public class DepartmentApiController {
     public ResponseEntity<?> get(@PathVariable Long id) {
         DepartmentResponseDto department = dataService.get(id);
         return ResponseEntity.ok(department);
-    }
-
-    @GetMapping(path = "/get/sub_departments")
-    public ResponseEntity<?> getSubDepartments(@RequestParam(name = "id") Long parentId) {
-        return ResponseEntity.ok(dataService.getSubDepartments(parentId));
     }
 
     @GetMapping(path = "/all")
