+++ hr-api/src/main/java/com/digdes/controllers/EmployeeApiController.java
@@ -8,7 +8,9 @@ import com.digdes.exceptions.EntityNotFoundException;
 import com.digdes.exceptions.EntityUpdateException;
 import com.digdes.services.EmployeeDataService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,23 +19,24 @@ import java.util.Optional;
 @RestController
 @RequestMapping(path = "/employees")
 public class EmployeeApiController{
-
-    //todo: constraints for bad request
     @Autowired
     private EmployeeDataService dataService;
 
     @PostMapping(path = "/create")
-    public ResponseEntity<?> create(@RequestBody EmployeeDto info) {
+    public ResponseEntity<?> create(@RequestBody EmployeeDto info, Authentication authentication) {
         if (info.getId() != null)
             return ResponseEntity.badRequest().body("Employee id must be null for create. Use update method");
         if (info.getFirstName() == null || info.getLastName() == null
                 || info.getBirthDate() == null || info.getGender() == null
                 || info.getEmail() == null || info.getDepartmentId() == null
-                || info.getPositionId() == null)
+                || info.getPositionId() == null || info.getHead() == null)
             return ResponseEntity.badRequest().body("Only employee patronymic field can be null. Check other fields");
         try{
-            EmployeeResponseDto result = dataService.create(info);
-            return ResponseEntity.ok(result);
+            if (dataService.checkUser(authentication.getName(), info.getDepartmentId())) {
+                EmployeeResponseDto result = dataService.create(info);
+                return ResponseEntity.ok(result);
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(String.format("User %s cannot moderate this employee", authentication.getName()));
         }
         catch (EntityNotFoundException | EntityCreateException e){
             return ResponseEntity.internalServerError().body(e.getMessage());
@@ -41,17 +44,20 @@ public class EmployeeApiController{
     }
 
     @PutMapping(path = "/update")
-    public ResponseEntity<?> update(@RequestBody EmployeeDto info) {
+    public ResponseEntity<?> update(@RequestBody EmployeeDto info, Authentication authentication) {
         if (info.getId() == null)
             return ResponseEntity.badRequest().body("Employee id must not be null for update. Use create method");
         if (info.getFirstName() == null && info.getLastName() == null
                 && info.getBirthDate() == null && info.getGender() == null
                 && info.getEmail() == null && info.getDepartmentId() == null
-                && info.getPositionId() == null && info.getPatronymic() == null)
+                && info.getPositionId() == null && info.getPatronymic() == null && info.getHead() == null)
             return ResponseEntity.badRequest().body("One of updating fields must not be null");
         try{
-            EmployeeResponseDto result = dataService.update(info);
-            return ResponseEntity.ok(result);
+            if (dataService.checkUser(authentication.getName(), info.getDepartmentId())) {
+                EmployeeResponseDto result = dataService.update(info);
+                return ResponseEntity.ok(result);
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(String.format("User %s cannot moderate this employee", authentication.getName()));
         }
         catch (EntityNotFoundException | EntityUpdateException e){
             return ResponseEntity.internalServerError().body(e.getMessage());
@@ -59,12 +65,15 @@ public class EmployeeApiController{
     }
 
     @DeleteMapping(path = "/delete")
-    public ResponseEntity<?> delete(@RequestBody EmployeeDto info) {
+    public ResponseEntity<?> delete(@RequestBody EmployeeDto info, Authentication authentication) {
         if (info.getId() == null)
             return ResponseEntity.badRequest().body("Employee id must not be null for delete");
         try {
-            Boolean result = dataService.delete(info);
-            return ResponseEntity.ok(result);
+            if (dataService.checkUser(authentication.getName(), info.getDepartmentId())) {
+                Boolean result = dataService.delete(info);
+                return ResponseEntity.ok(result);
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(String.format("User %s cannot moderate this employee", authentication.getName()));
         }
         catch (EntityNotFoundException | EntityDeleteException e){
             return ResponseEntity.internalServerError().body(e.getMessage());
@@ -77,7 +86,7 @@ public class EmployeeApiController{
                 && searchRequest.getBirthDate() == null && searchRequest.getGender() == null
                 && searchRequest.getEmail() == null && searchRequest.getDepartmentId() == null
                 && searchRequest.getPositionId() == null && searchRequest.getPatronymic() == null
-                && searchRequest.getId() == null)
+                && searchRequest.getId() == null && searchRequest.getHead() == null)
             return ResponseEntity.badRequest().body("One of fields must not be null");
         try {
             List<EmployeeResponseDto> result = dataService.find(searchRequest);
